Update energyGain.py
Update the power ratio and percent change metrics to make sure they worked correctly when you don't want to use test turbines (previously had only updated the AEP method), added a boolean attribute to the constructor for whether you want to include test turbines in any calculations for the energyGain object, and started working on the line plot but did not finish. Currently the axes are distorted.