Fixed huge bug in bootstrapEstimate
When nDim=1, bootstrapEstimate's default was to use aepGain as the default metric. But the way it called aepGain did not give it the proper inputs. I think this is a bigger issue with 1/2 dimensional metric methods, so as a quick fix I just made it call percentPowerGain instead. Ideally, the user can pass in the metricMethod instead of having one hard coded into the method